server/requirements.txt
aniso8601==7.0.0
blinker==1.8.2
click==8.1.7
Flask==3.0.3
Flask-RESTful==0.3.10
Flask-Cors==5.0.1
strawberry-graphql==0.283.3
importlib_metadata==8.2.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
promise==2.3
pytz==2024.1
Rx==1.6.3
six==1.16.0
Werkzeug==3.0.3
zipp==3.19.2

# Testing dependencies
pytest==7.4.0
pytest-cov==4.1.0
pytest-mock==3.10.0
httpx==0.24.0
freezegun==1.2.2
factory_boy==3.2.1
Faker==18.9.1
types-factory-boy==0.4.1
